// src/Dogger/wwwroot/src/api/openapi/models/PlanResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    PullDogPlanResponse,
    PullDogPlanResponseFromJSON,
    PullDogPlanResponseFromJSONTyped,
    PullDogPlanResponseToJSON,
} from './';

/**
 * 
 * @export
 * @interface PlanResponse
 */
export interface PlanResponse {
    /**
     * 
     * @type {string}
     * @memberof PlanResponse
     */
    id?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    priceInHundreds?: number;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    ramSizeInMegabytes?: number;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    cpuCount?: number;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    power?: number;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    diskSizeInGigabytes?: number;
    /**
     * 
     * @type {number}
     * @memberof PlanResponse
     */
    transferPerMonthInGigabytes?: number;
    /**
     * 
     * @type {Array<PullDogPlanResponse>}
     * @memberof PlanResponse
     */
    pullDogPlans?: Array<PullDogPlanResponse> | null;
}

export function PlanResponseFromJSON(json: any): PlanResponse {
    return PlanResponseFromJSONTyped(json, false);
}

export function PlanResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PlanResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'priceInHundreds': !exists(json, 'priceInHundreds') ? undefined : json['priceInHundreds'],
        'ramSizeInMegabytes': !exists(json, 'ramSizeInMegabytes') ? undefined : json['ramSizeInMegabytes'],
        'cpuCount': !exists(json, 'cpuCount') ? undefined : json['cpuCount'],
        'power': !exists(json, 'power') ? undefined : json['power'],
        'diskSizeInGigabytes': !exists(json, 'diskSizeInGigabytes') ? undefined : json['diskSizeInGigabytes'],
        'transferPerMonthInGigabytes': !exists(json, 'transferPerMonthInGigabytes') ? undefined : json['transferPerMonthInGigabytes'],
        'pullDogPlans': !exists(json, 'pullDogPlans') ? undefined : (json['pullDogPlans'] === null ? null : (json['pullDogPlans'] as Array<any>).map(PullDogPlanResponseFromJSON)),
    };
}

export function PlanResponseToJSON(value?: PlanResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'priceInHundreds': value.priceInHundreds,
        'ramSizeInMegabytes': value.ramSizeInMegabytes,
        'cpuCount': value.cpuCount,
        'power': value.power,
        'diskSizeInGigabytes': value.diskSizeInGigabytes,
        'transferPerMonthInGigabytes': value.transferPerMonthInGigabytes,
        'pullDogPlans': value.pullDogPlans === undefined ? undefined : (value.pullDogPlans === null ? null : (value.pullDogPlans as Array<any>).map(PullDogPlanResponseToJSON)),
    };
}