// src/Dogger/wwwroot/src/api/openapi/models/PullDogPlanResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * General
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PullDogPlanResponse
 */
export interface PullDogPlanResponse {
    /**
     * 
     * @type {string}
     * @memberof PullDogPlanResponse
     */
    id?: string | null;
    /**
     * 
     * @type {number}
     * @memberof PullDogPlanResponse
     */
    priceInHundreds?: number;
    /**
     * 
     * @type {number}
     * @memberof PullDogPlanResponse
     */
    poolSize?: number;
}

export function PullDogPlanResponseFromJSON(json: any): PullDogPlanResponse {
    return PullDogPlanResponseFromJSONTyped(json, false);
}

export function PullDogPlanResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PullDogPlanResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'priceInHundreds': !exists(json, 'priceInHundreds') ? undefined : json['priceInHundreds'],
        'poolSize': !exists(json, 'poolSize') ? undefined : json['poolSize'],
    };
}

export function PullDogPlanResponseToJSON(value?: PullDogPlanResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'priceInHundreds': value.priceInHundreds,
        'poolSize': value.poolSize,
    };
}